refactor: move `get_fullscreen_elemetn` to `fullscreen/lib.rs`
